Renamed LayoutManager2D to Layout; Renamed PlotLayout to EdgeLayout; Added new StackedLayout class; Plot now derives from DrawableContainer instead of JPanel; Added separate class PlotPnael for Swing integration; Renamed several methods of DrawableContainer;

+++ src/openjchart/plots/XYPlot.java
@@ -189,9 +189,8 @@ public class XYPlot extends Plot {
 	}
 
 	@Override
-	public void setBounds(int x, int y, int width, int height) {
+	public void setBounds(double x, double y, double width, double height) {
 		super.setBounds(x, y, width, height);
-
 		layoutAxes();
 	}
